refactor: snackbar
