--- packages/web/src/hooks/common/use-notice.ts
@@ -1,6 +1,0 @@
-import { useContext } from "react";
-import { NoticeContext } from "@context/NoticeContext";
-
-export const useNotice = () => {
-  return useContext(NoticeContext);
-};
