+++ packages/web/src/providers/snackbar-provider/index.ts
@@ -1,0 +1,5 @@
+import SnackbarProvider, { SnackbarContext } from "./SnackbarProvider";
+
+export * from "./type";
+export { SnackbarContext };
+export default SnackbarProvider;
