+++ packages/web/src/providers/snackbar-provider/snackbar/contents/type.ts
@@ -1,0 +1,12 @@
+export type SnackbarType =
+  | "success"
+  | "error"
+  | "pending"
+  | "withdraw-success"
+  | "withdraw-error";
+  
+export interface SnackbarContent {
+  title?: string;
+  description?: string;
+  txHash?: string;
+}
