+++ packages/web/src/providers/snackbar-provider/type.ts
@@ -1,0 +1,10 @@
+import { SnackbarContent, SnackbarType } from "./snackbar";
+
+export interface SnackbarOptions {
+  type: SnackbarType;
+  id: number;
+  timeout: number;
+  closeable?: boolean;
+  onClose?: () => void;
+  data?: SnackbarContent;
+}
